refactor(ErrorBoundary): Extract Error presentational component
