+++ src/components/ErrorBoundary/Error.styl
