+++ src/components/ErrorBoundary/index.js
@@ -1,0 +1,5 @@
+import ErrorBoundary from './ErrorBoundary'
+import Error from './Error'
+
+export default ErrorBoundary
+export { Error }
